ENG-308: Implement scan secrets GitHub action (#61)
## Description

- Fixes #35 
- Informs #46 


## Context

After investigation and review by the team it was decided that gitleaks
was a good fit for the requirement to guard against storing secrets in
the repo

## Type of changes

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [x] I have updated the documentation accordingly
- [x] This PR is a result of pair or mob programming

At this stage we did manual testing and added the need for automated
tests to the TODO
---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.

---------

Co-authored-by: Jonathan Pearce <j.pearce1@nhs.net>
Co-authored-by: Tamara Goldschmidt <tamaragoldschmidt@users.noreply.github.com>
Co-authored-by: Tim Rickwood <tim.rickwood@nhs.net>