Add a basic test infrastructure (#86)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

This PR adds `make` tasks for running tests.

## Context

[This
issue](https://github.com/NHSDigital/software-engineering-quality-framework/issues/297)
says that we're out of alignment with wider policy in the sense that
we're applying stricter rules than government policy requires.

There are two ways to solve this: we can either loosen our rules, or
make them easier to comply with. From the point of view of the test
statuses we require in the quality framework
[here](https://github.com/NHSDigital/software-engineering-quality-framework/blob/main/quality-checks.md?plain=1#L66-L86),
this does the latter.

Specifically it allows the user of the template to put scripts into
`scripts/test` for running the various types of test required by the
`software-engineering-quality-framework` which, until they are
implemented, will output a warning to the console (but do not fail).

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [ ] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [ ] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ ] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.

---------

Signed-off-by: regularfry <alex@blackkettle.org>