Update scan dependencies action (#94)
## Description

- Include fully working and documented SBOM and vulnerabilities
reporting functionality
- Upload both SBOM and vulnerabilities reports as zip archives
- Get `datetime` from the workflow for traceability 
- Add `IDP_*` variables to configure upload of the SBOM and
vulnerabilities reports
- Use defensive coding mechanisms to ensure the reporting functionality
will work 'as-is' without failing the pipeline when dropped into a new
repository, even when the `IDP_*` variables are not set
- Relocate shell scripts to the `./scripts/reports` directory
- Introduce a Git pre-commit hook
- Update Docker image version used by Syft and Grype

## Context

This change ensures the ease of adoption of the SBOM and vulnerabilities
reporting functionality.

## Type of changes

- [x] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [x] I have added tests to cover my changes
- [x] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.