Change how check-file-format.sh checks changed files (#130)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

Prior to this patch, the `git diff` command used to identify which files
to check for editorconfig compliance caught too many files, and did the
expensive part of the check inside the per-file loop.

By lucky chance, the editorconfig-checker container image has a `git`
binary installed, so we can move the entire loop inside a single
container invocation.

Fixes #129.

## Context

It makes the editorconfig check, which is called in a pre-commit hook,
fast.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [x] Bug fix (non-breaking change which fixes an issue)