Remove commas after the word "please"
A comma before "please" at the end of a sentence is often necessary, but a
comma after "please" rarely is.  The exception would be in sentences that have
an interposed "please" somewhere in the middle, like "Make sure, please, to
switch off the nuclear reactor before attempting to clean the reaction vessel."
But I don't think we've got any sentences like that hanging around.