Refactor shell scripts (#141)
## Description

This is the second part of the refactoring effort to ensure that the
shell scripts:

- Follow a consistent convention,
- Pass the ShellCheck linting test (please run `make
shellscript-lint-all`), and
- Work either with the pre-installed CLI tooling or run commands using
Docker containers.

## Context

Information for a reviewer on how to test the scripts with the
`scan-secrets.sh` as an example:

Install the CLI tool

```
# On macOS
brew install gitleaks
# On Ubuntu
apt install gitleaks
```

Run a test

```
./scripts/githooks/scan-secrets.sh
VERBOSE=1 ./scripts/githooks/scan-secrets.sh # Use the gitleaks CLI tool if installed
VERBOSE=1 FORCE_USE_DOCKER=1 ./scripts/githooks/scan-secrets.sh # Use Docker
```

## Type of changes

- [x] Refactoring (non-breaking change)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.