Add the _install-dependencies task and call it from make config (#144)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

This adds the `make _install-dependencies` task, which installs
everything from the `.tool-versions` file.

## Context

This avoids the need to add a new `make` task for each dependency,
instead treating the `.tool-versions` file as canonical. Specifically it
avoids the need for the user to add a `make _install-python` task in
[the
tutorial](https://github.com/nhs-england-tools/repository-template/pull/131).

For discussion: do we still need `make _install-terraform` if we have
this? I don't think we do.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [ ] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [ ] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [ ] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ ] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.