Add Vale pre-commit Git hook (#146)
## Description

This change adds a Vale pre-commit Git hook to the repository template.
Additionally, the following refactoring tasks were done:

- Use Vale natively, if available, otherwise fall back to Docker.
- Maintain a consistent naming convention for GitHub actions and
pre-commit Git hooks.

## Context

Vale is a command-line tool that functions as a syntax-aware linter for
prose. It's designed to help writers adhere to a specific writing style
or set of guidelines. We automatically check the English usage in our
documentation.

## Type of changes

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [x] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [x] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.