Reworked docker docs (#155)
<!-- markdownlint-disable-next-line first-line-heading -->
## Description

Minor (but important!) changes to how docker is run, `DOCKER_IMAGE`
handling, and some docs.

## Context

This change wraps up a few different changes I needed to make when using
the template.

### Docker context


[Here](https://github.com/nhs-england-tools/repository-template/compare/alyo12-docker-docs?expand=1#diff-dddd47b45acd050e3b06771b95c68eee8ebcbfcf8a019b3f0186725a18cfcab6L33)
we were `cd`ing into the `infrastructure/images/whatever` directory
before running Docker. That restricts the docker context to only that
directory and below. You can't `COPY ../../. .` or whatever because
that's treated as a security problem.

This meant that you couldn't put your application code into a container
inside the `infrastructure/images` structure. That's fine if you've got
a single application container, and if you're happy with it living at
the root of the repo, but as soon as you've got more than one it gets a
bit messy (and I've never liked things that clutter the repo root at the
best of times).

The change I've made gets rid of the `cd` in and out; `docker` is
*always* run from the repo root. That means paths in `Dockerfile`s are
relative to the root, not to themselves; same with `.dockerignore`
(which I've made optional so that it doesn't break if there isn't one).

### `$DOCKER_IMAGE` variable handling

Minor tweak but a good quality of life improvement here. If you specify
`$DOCKER_IMAGE` to the `make docker-*` tasks but *not* `$dir` then it
defaults `$dir` to `infrastructure/images/$DOCKER_IMAGE`. That means the
`make` invocation, the filesystem, and the docker image list are all
synchronised to the same name, by convention.

### Updated docker usage docs

The example in the existing docs was specific to repackaging a third
party tool, but gave no guidance about how to package your own
application code. I've rewritten the `Quick start` section to show
packaging a trivial python script. I'm a little worried that the docs
are a bit repetitive now, but I don't think it's worth reworking right
now.

I've also removed the python example code. It was a good start, but in
practice having the tutorial docs is better because what I want to know
as a new user is *where to put my own stuff*, and the example wasn't a
huge help there and in fact was a bit of a confusion.

## Type of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply. -->

- [x] Refactoring (non-breaking change)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would change existing
functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply. -->

- [ ] I am familiar with the [contributing
guidelines](../docs/CONTRIBUTING.md)
- [ ] I have followed the code style of the project
- [ ] I have added tests to cover my changes
- [ ] I have updated the documentation accordingly
- [ ] This PR is a result of pair or mob programming

---

## Sensitive Information Declaration

To ensure the utmost confidentiality and protect your and others
privacy, we kindly ask you to NOT including [PII (Personal Identifiable
Information) / PID (Personal Identifiable
Data)](https://digital.nhs.uk/data-and-information/keeping-data-safe-and-benefitting-the-public)
or any other sensitive data in this PR (Pull Request) and the codebase
changes. We will remove any PR that do contain any sensitive
information. We really appreciate your cooperation in this matter.

- [ ] I confirm that neither PII/PID nor sensitive data are included in
this PR and the codebase changes.